UNSTABLE Moved functions to hypervisor interface
