+++ pkg/hypervisor/BUILD.bazel
@@ -7,6 +7,8 @@ go_library(
         "hypervisor.go",
         "qemu-hypervisor.go",
     ],
+    deps = [
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/hypervisor",
     visibility = ["//visibility:public"],
 )
