+++ pkg/virt-launcher/virtwrap/util/libvirt_helper.go
@@ -1,24 +1,11 @@
 package util
 
 import (
-	"bufio"
 	"encoding/base64"
 	"encoding/xml"
-	"errors"
 	"fmt"
-	"io"
-	"os"
-	"os/exec"
-	"path"
-	"path/filepath"
-	"strconv"
 	"strings"
-	"syscall"
-	"time"
 
-	"kubevirt.io/kubevirt/pkg/virt-controller/services"
-
-	"golang.org/x/sys/unix"
 	k8sv1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"libvirt.org/go/libvirt"
@@ -29,8 +16,6 @@ import (
 	"kubevirt.io/client-go/log"
 
 	"kubevirt.io/kubevirt/pkg/hooks"
-	"kubevirt.io/kubevirt/pkg/util"
-	putil "kubevirt.io/kubevirt/pkg/util"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/cli"
 )
@@ -97,118 +82,11 @@ var getHookManager = hooks.GetManager
 
 // This interface exposes functions used by virt-launcher to start and configure libvirt. The implementation varies for different hypervisors.
 type LibvirtWrapper interface {
-	// Setup libvirt for hosting the virtual machine. This function is called during the startup of the virt-launcher.
-	SetupLibvirt(customLogFilters *string) (err error)
+
 	// Return a list of potential prefixes of the specific hypervisor's process, e.g., qemu-system or cloud-hypervisor
 	GetHypervisorCommandPrefix() []string
-	// Start the libvirt daemon, either in modular mode or monolithic mode
-	StartHypervisorDaemon(stopChan chan struct{})
 	// Start the virtlogd daemon, which is used to capture logs from the hypervisor
 	StartVirtlog(stopChan chan struct{}, domainName string)
-	// Returns true if the libvirt process(es) should be run as root user
-	root() bool
-	// Return the URI to connect to libvirt and the user with which to connect to libvirt
-	GetLibvirtUriAndUser() (string, string)
-	// Return the directory where libvirt stores the PID files of the hypervisor processes
-	GetPidDir() string
-	// Return the modular libvirt daemon for the hypervisor
-	GetModularDaemonName() string
-	// Return the VMM name
-	GetVmm() string
-}
-
-type QemuLibvirtWrapper struct {
-	user uint32
-}
-
-func (q *QemuLibvirtWrapper) StartHypervisorDaemon(stopChan chan struct{}) {
-	startModularLibvirtDaemon(q, stopChan)
-}
-
-func (q *QemuLibvirtWrapper) GetPidDir() string {
-	if q.root() {
-		return "/run/libvirt/qemu"
-	} else {
-		return "/run/libvirt/qemu/run"
-	}
-}
-
-func (q *QemuLibvirtWrapper) GetHypervisorCommandPrefix() []string {
-	// give qemu some time to shut down in case it survived virt-handler
-	// Most of the time we call `qemu-system=* binaries, but qemu-system-* packages
-	// are not everywhere available where libvirt and qemu are. There we usually call qemu-kvm
-	// which resides in /usr/libexec/qemu-kvm
-	return []string{"qemu-system", "qemu-kvm"}
-}
-
-func (q *QemuLibvirtWrapper) GetLibvirtUriAndUser() (string, string) {
-	libvirtUri := "qemu:///system"
-	user := ""
-	if !q.root() {
-		user = putil.NonRootUserString
-		libvirtUri = "qemu+unix:///session?socket=/var/run/libvirt/virtqemud-sock"
-	}
-	return libvirtUri, user
-}
-
-func (q *QemuLibvirtWrapper) GetModularDaemonName() string {
-	return "virtqemud"
-}
-
-func (q *QemuLibvirtWrapper) GetVmm() string {
-	return "qemu"
-}
-
-func (c *CloudHypervisorLibvirtWrapper) GetPidDir() string {
-	return "/run/libvirt/ch"
-}
-
-func (c *CloudHypervisorLibvirtWrapper) GetLibvirtUriAndUser() (string, string) {
-	return "ch:///system", ""
-}
-
-func (c *CloudHypervisorLibvirtWrapper) GetHypervisorCommandPrefix() []string {
-	return []string{"cloud-hypervisor"}
-}
-
-func (c *CloudHypervisorLibvirtWrapper) StartHypervisorDaemon(stopChan chan struct{}) {
-	startModularLibvirtDaemon(c, stopChan)
-}
-
-func (c *CloudHypervisorLibvirtWrapper) GetVmm() string {
-	return "ch"
-}
-
-type CloudHypervisorLibvirtWrapper struct {
-	user uint32
-}
-
-func (l *CloudHypervisorLibvirtWrapper) SetupLibvirt(customLogFilters *string) (err error) {
-	return setupLibvirt(l, customLogFilters, false)
-}
-
-func (c *CloudHypervisorLibvirtWrapper) GetModularDaemonName() string {
-	return "virtchd"
-}
-
-func NewLibvirtWrapper(nonRoot bool, vmm string) LibvirtWrapper {
-	if vmm == "ch" {
-		return &CloudHypervisorLibvirtWrapper{
-			user: util.RootUser,
-		}
-	} else if vmm == "qemu" {
-		if nonRoot {
-			return &QemuLibvirtWrapper{
-				user: util.NonRootUID,
-			}
-		}
-		return &QemuLibvirtWrapper{
-			user: util.RootUser,
-		}
-	} else {
-		return nil
-	}
-
 }
 
 func ConvState(status libvirt.DomainState) api.LifeCycle {
@@ -311,192 +189,6 @@ func GetDomainSpecWithFlags(dom cli.VirDomain, flags libvirt.DomainXMLFlags) (*a
 	return domain, nil
 }
 
-func startModularLibvirtDaemon(l LibvirtWrapper, stopChan chan struct{}) {
-	// we spawn libvirt from virt-launcher in order to ensure the virtqemud+qemu process
-	// doesn't exit until virt-launcher is ready for it to. Virt-launcher traps signals
-	// to perform special shutdown logic. These processes need to live in the same
-	// container.
-
-	modularDaemonName := l.GetModularDaemonName()
-
-	go func() {
-		for {
-			exitChan := make(chan struct{})
-			args := []string{"-f", fmt.Sprintf("/var/run/libvirt/%s.conf", modularDaemonName)}
-			cmd := exec.Command(fmt.Sprintf("/usr/sbin/%s", modularDaemonName), args...)
-			if !l.root() {
-				cmd.SysProcAttr = &syscall.SysProcAttr{
-					AmbientCaps: []uintptr{unix.CAP_NET_BIND_SERVICE},
-				}
-			}
-
-			// connect libvirt's stderr to our own stdout in order to see the logs in the container logs
-			reader, err := cmd.StderrPipe()
-			if err != nil {
-				log.Log.Reason(err).Error(fmt.Sprintf("failed to start %s", modularDaemonName))
-				panic(err)
-			}
-
-			go func() {
-				scanner := bufio.NewScanner(reader)
-				scanner.Buffer(make([]byte, 1024), 512*1024)
-				for scanner.Scan() {
-					log.LogLibvirtLogLine(log.Log, scanner.Text())
-				}
-
-				if err := scanner.Err(); err != nil {
-					log.Log.Reason(err).Error("failed to read libvirt logs")
-				}
-			}()
-
-			err = cmd.Start()
-			if err != nil {
-				log.Log.Reason(err).Error(fmt.Sprintf("failed to start %s", modularDaemonName))
-				panic(err)
-			}
-
-			go func() {
-				defer close(exitChan)
-				cmd.Wait()
-			}()
-
-			select {
-			case <-stopChan:
-				cmd.Process.Kill()
-				return
-			case <-exitChan:
-				log.Log.Errorf(fmt.Sprintf("%s exited, restarting", modularDaemonName))
-			}
-
-			// this sleep is to avoid consuming all resources in the
-			// event of a virtqemud crash loop.
-			time.Sleep(time.Second)
-		}
-	}()
-}
-
-func startVirtlogdLogging(virtlogdBinaryPath string, stopChan chan struct{}, domainName string, nonRoot bool) {
-	for {
-		cmd := exec.Command(virtlogdBinaryPath, "-f", "/etc/libvirt/virtlogd.conf")
-
-		exitChan := make(chan struct{})
-
-		err := cmd.Start()
-		if err != nil {
-			log.Log.Reason(err).Error("failed to start virtlogd")
-			panic(err)
-		}
-
-		go func() {
-			logfile := fmt.Sprintf("/var/log/libvirt/qemu/%s.log", domainName)
-			if nonRoot {
-				logfile = filepath.Join("/var", "run", "kubevirt-private", "libvirt", "qemu", "log", fmt.Sprintf("%s.log", domainName))
-			}
-
-			// It can take a few seconds to the log file to be created
-			for {
-				_, err = os.Stat(logfile)
-				if !errors.Is(err, os.ErrNotExist) {
-					break
-				}
-				time.Sleep(time.Second)
-			}
-			// #nosec No risk for path injection. logfile has a static basedir
-			file, err := os.Open(logfile)
-			if err != nil {
-				errMsg := fmt.Sprintf("failed to open logfile in path: \"%s\"", logfile)
-				log.Log.Reason(err).Error(errMsg)
-				return
-			}
-			defer util.CloseIOAndCheckErr(file, nil)
-
-			scanner := bufio.NewScanner(file)
-			scanner.Buffer(make([]byte, 1024), 512*1024)
-			for scanner.Scan() {
-				log.LogQemuLogLine(log.Log, scanner.Text())
-			}
-
-			if err := scanner.Err(); err != nil {
-				log.Log.Reason(err).Error("failed to read virtlogd logs")
-			}
-		}()
-
-		go func() {
-			defer close(exitChan)
-			_ = cmd.Wait()
-		}()
-
-		select {
-		case <-stopChan:
-			_ = cmd.Process.Kill()
-			return
-		case <-exitChan:
-			log.Log.Errorf("virtlogd exited, restarting")
-		}
-
-		// this sleep is to avoid consuming all resources in the
-		// event of a virtlogd crash loop.
-		time.Sleep(time.Second)
-	}
-}
-
-func startQEMUSeaBiosLogging(stopChan chan struct{}) {
-	const QEMUSeaBiosDebugPipeMode uint32 = 0666
-	const logLinePrefix = "[SeaBios]:"
-
-	err := syscall.Mkfifo(QEMUSeaBiosDebugPipe, QEMUSeaBiosDebugPipeMode)
-	if err != nil {
-		log.Log.Reason(err).Error(fmt.Sprintf("%s failed creating a pipe for sea bios debug logs", logLinePrefix))
-		return
-	}
-
-	// Chmod is needed since umask is 0018. Therefore Mkfifo does not actually create a pipe with proper permissions.
-	err = syscall.Chmod(QEMUSeaBiosDebugPipe, QEMUSeaBiosDebugPipeMode)
-	if err != nil {
-		log.Log.Reason(err).Error(fmt.Sprintf("%s failed executing chmod on pipe for sea bios debug logs.", logLinePrefix))
-		return
-	}
-
-	QEMUPipe, err := os.OpenFile(QEMUSeaBiosDebugPipe, os.O_RDONLY, 0604)
-
-	if err != nil {
-		log.Log.Reason(err).Error(fmt.Sprintf("%s failed to open %s", logLinePrefix, QEMUSeaBiosDebugPipe))
-		return
-	}
-	defer QEMUPipe.Close()
-
-	scanner := bufio.NewScanner(QEMUPipe)
-	for {
-		for scanner.Scan() {
-			logLine := fmt.Sprintf("%s %s", logLinePrefix, scanner.Text())
-
-			log.LogQemuLogLine(log.Log, logLine)
-
-			select {
-			case <-stopChan:
-				return
-			default:
-			}
-		}
-
-		if err := scanner.Err(); err != nil {
-			log.Log.Reason(err).Error(fmt.Sprintf("%s reader failed with an error", logLinePrefix))
-			return
-		}
-
-		log.Log.Errorf(fmt.Sprintf("%s exited, restarting", logLinePrefix))
-	}
-}
-
-func (l *QemuLibvirtWrapper) StartVirtlog(stopChan chan struct{}, domainName string) {
-	go startVirtlogdLogging("/usr/sbin/virtlogd", stopChan, domainName, l.user != util.RootUser)
-	go startQEMUSeaBiosLogging(stopChan)
-}
-
-func (l *CloudHypervisorLibvirtWrapper) StartVirtlog(stopChan chan struct{}, domainName string) { // TODO No need for this function when the path to virtlogd has been made consistent with QEMU
-	go startVirtlogdLogging("/usr/sbin/virtlogd", stopChan, domainName, l.user != util.RootUser)
-}
-
 // returns the namespace and name that is encoded in the
 // domain name.
 func SplitVMINamespaceKey(domainName string) (namespace, name string) {
@@ -537,165 +229,4 @@ func NewDomainFromName(name string, vmiUID types.UID) *api.Domain {
 	domain.Spec.Metadata.KubeVirt.UID = vmiUID
 	domain.GetObjectMeta().SetUID(domain.Spec.Metadata.KubeVirt.UID)
 	return domain
-}
-
-func configureVmmConf(vmmConfFilename string) (err error) {
-	vmmConf, err := os.OpenFile(vmmConfFilename, os.O_APPEND|os.O_WRONLY, 0644)
-	if err != nil {
-		return err
-	}
-	defer util.CloseIOAndCheckErr(vmmConf, &err)
-
-	// If hugepages exist, tell libvirt about them
-	_, err = os.Stat("/dev/hugepages")
-	if err == nil {
-		_, err = vmmConf.WriteString("hugetlbfs_mount = \"/dev/hugepages\"\n")
-	} else if !errors.Is(err, os.ErrNotExist) {
-		return err
-	}
-
-	if envVarValue, ok := os.LookupEnv("VIRTIOFSD_DEBUG_LOGS"); ok && (envVarValue == "1") {
-		_, err = vmmConf.WriteString("virtiofsd_debug = 1\n")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func copyFile(from, to string) error {
-	f, err := os.OpenFile(from, os.O_RDONLY, 0644)
-	if err != nil {
-		return err
-	}
-	defer util.CloseIOAndCheckErr(f, &err)
-	newFile, err := os.OpenFile(to, os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return err
-	}
-	defer util.CloseIOAndCheckErr(newFile, &err)
-
-	_, err = io.Copy(newFile, f)
-	return err
-}
-
-func setupLibvirt(l LibvirtWrapper, customLogFilters *string, shouldConfigureVmmConf bool) (err error) {
-	if shouldConfigureVmmConf {
-		vmmConfPath := fmt.Sprintf(vmmConfPathPattern, l.GetVmm())
-		runtimeVmmConfPath := vmmConfPath
-		if !l.root() {
-			runtimeVmmConfPath = fmt.Sprintf(vmmNonRootConfPathPattern, l.GetVmm())
-
-			if err := os.MkdirAll(libvirtHomePath, 0755); err != nil {
-				return err
-			}
-			if err := copyFile(vmmConfPath, runtimeVmmConfPath); err != nil {
-				return err
-			}
-		}
-
-		if err := configureVmmConf(runtimeVmmConfPath); err != nil {
-			return err
-		}
-	}
-
-	runtimeVmmDaemonConfPath := path.Join(libvirtRuntimePath, fmt.Sprintf("%s.conf", l.GetModularDaemonName()))
-	vmmModularDaemonConfPath := fmt.Sprintf(vmmModularDaemonConfPathPattern, l.GetModularDaemonName())
-	if err := copyFile(vmmModularDaemonConfPath, runtimeVmmDaemonConfPath); err != nil {
-		return err
-	}
-
-	var libvirtLogVerbosityEnvVar *string
-	if envVarValue, envVarDefined := os.LookupEnv(services.ENV_VAR_VIRT_LAUNCHER_LOG_VERBOSITY); envVarDefined {
-		libvirtLogVerbosityEnvVar = &envVarValue
-	}
-	_, libvirtDebugLogsEnvVarDefined := os.LookupEnv(services.ENV_VAR_LIBVIRT_DEBUG_LOGS)
-
-	if logFilters, enableDebugLogs := getLibvirtLogFilters(customLogFilters, libvirtLogVerbosityEnvVar, libvirtDebugLogsEnvVarDefined); enableDebugLogs {
-		virtqemudConf, err := os.OpenFile(runtimeVmmDaemonConfPath, os.O_APPEND|os.O_WRONLY, 0644)
-		if err != nil {
-			return err
-		}
-		defer util.CloseIOAndCheckErr(virtqemudConf, &err)
-
-		log.Log.Infof("Enabling libvirt log filters: %s", logFilters)
-		_, err = virtqemudConf.WriteString(fmt.Sprintf("log_filters=\"%s\"\n", logFilters))
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (l *QemuLibvirtWrapper) SetupLibvirt(customLogFilters *string) (err error) {
-	return setupLibvirt(l, customLogFilters, true)
-}
-
-// getLibvirtLogFilters returns libvirt debug log filters that should be enabled if enableDebugLogs is true.
-// The decision is based on the following logic:
-//   - If custom log filters are defined - they should be enabled and used.
-//   - If verbosity is defined and beyond threshold then debug logs would be enabled and determined by verbosity level
-//   - If verbosity level is below threshold but debug logs environment variable is defined, debug logs would be enabled
-//     and set to the highest verbosity level.
-//   - If verbosity level is below threshold and debug logs environment variable is not defined - debug logs are disabled.
-func getLibvirtLogFilters(customLogFilters, libvirtLogVerbosityEnvVar *string, libvirtDebugLogsEnvVarDefined bool) (logFilters string, enableDebugLogs bool) {
-
-	if customLogFilters != nil && *customLogFilters != "" {
-		return *customLogFilters, true
-	}
-
-	var libvirtLogVerbosity int
-	var err error
-
-	if libvirtLogVerbosityEnvVar != nil {
-		libvirtLogVerbosity, err = strconv.Atoi(*libvirtLogVerbosityEnvVar)
-		if err != nil {
-			log.Log.Infof("cannot apply %s value %s - must be a number", services.ENV_VAR_VIRT_LAUNCHER_LOG_VERBOSITY, *libvirtLogVerbosityEnvVar)
-			libvirtLogVerbosity = -1
-		}
-	} else {
-		libvirtLogVerbosity = -1
-	}
-
-	const verbosityThreshold = services.EXT_LOG_VERBOSITY_THRESHOLD
-
-	if libvirtLogVerbosity < verbosityThreshold {
-		if libvirtDebugLogsEnvVarDefined {
-			libvirtLogVerbosity = verbosityThreshold + 5
-		} else {
-			return "", false
-		}
-	}
-
-	// Higher log level means higher verbosity
-	const logsLevel4 = "3:remote 4:event 3:util.json 3:util.object 3:util.dbus 3:util.netlink 3:node_device 3:rpc 3:access"
-	const logsLevel3 = logsLevel4 + " 3:util.threadjob 3:cpu.cpu"
-	const logsLevel2 = logsLevel3 + " 3:qemu.qemu_monitor"
-	const logsLevel1 = logsLevel2 + " 3:qemu.qemu_monitor_json 3:conf.domain_addr"
-	const allowAllOtherCategories = " 1:*"
-
-	switch libvirtLogVerbosity {
-	case verbosityThreshold:
-		logFilters = logsLevel1
-	case verbosityThreshold + 1:
-		logFilters = logsLevel2
-	case verbosityThreshold + 2:
-		logFilters = logsLevel3
-	case verbosityThreshold + 3:
-		logFilters = logsLevel4
-	default:
-		logFilters = logsLevel4
-	}
-
-	return logFilters + allowAllOtherCategories, true
-}
-
-func (l *QemuLibvirtWrapper) root() bool {
-	return l.user == util.RootUser
-}
-
-func (l *CloudHypervisorLibvirtWrapper) root() bool {
-	return l.user == util.RootUser
 }
